Make NotAvailable.tsx reusable
